services/management-plane/pyproject.toml
[project]
name = "management-plane"
version = "0.1.0"
description = "Agent Hub Cloud Management Plane"
requires-python = ">=3.11"
dependencies = [
    "starlette>=0.40",
    "uvicorn>=0.30",
    "aiosqlite>=0.20",
    "bcrypt>=4.0",
    "pyjwt>=2.8",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/management"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
